requirements.txt
customtkinter==5.2.2
lark==1.2.2
numpy==1.26.4
pillow==10.2.0
watchdog==6.0.0
